pyproject.toml
[tool.poetry]
name = "team5"
version = "0.1.0"
description = ""
authors = ["mr.xed <25272822+JouharBirakdar223787@users.noreply.github.com>",
    "bobo_bobkata <112721901+BorislavNachev220472@users.noreply.github.com>",
    "lala <112754523+neildaniel221270@users.noreply.github.com>",
    "andrea <112705400+AndreaTosheva220666@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9.7,<=3.11.0"
pandas = "1.5.2"
numpy = "^1.24.3"
beautifulsoup4 = "^4.12.2"
requests = "^2.30.0"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.14"
osmnx = "^1.3.0"
geopandas = "^0.13.0"
scikit-learn = "^1.2.2"
topojson = "^1.5"
geojson = "^3.0.1"
statsmodels = "^0.14.0"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
sphinx = "^7.0.1"
streamlit = "^1.23.1"
python-dotenv = "^1.0.0"
plotly = "^5.14.1"
coverage = "^7.2.7"
matplotlib = "^3.7.1"
streamlit-option-menu = "^0.3.6"
streamlit-extras = "^0.2.7"
leafmap = "^0.21.3"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[tool.poetry.scripts]
configure_env = { callable = "src.main:download_and_preprocess_data", extras = ["all"] }
preprocess_data = { callable = "src.main:preprocess_downloaded_raw_data", extras = ["all"] }
sync_model = { callable = "src.main:train_model_on_new_data", extras = ["all"] }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
